+++ testproject/src/com/move/dependent_ns1.clj
@@ -5,6 +5,22 @@
             TypeToBeMoved RecordToBeMoved]))
 
 (defn- use-some-publics []
+
+  (com.move.ns-to-be-moved/fn-to-be-moved :_)
+
+  'com.move.ns-to-be-moved/fn-to-be-moved
+
+  #'com.move.ns-to-be-moved/fn-to-be-moved
+
+  ^:com.move.ns-to-be-moved/fn-to-be-moved []
+
+  #:com.move.ns-to-be-moved {:a :b}
+
+  (let [^com.move.ns_to_be_moved.TypeToBeMoved x (fn-to-be-moved :_)]
+    x)
+
+  (com.move.ns_to_be_moved.TypeToBeMoved. :ok)
+
   (macro-to-be-moved
    (fn-to-be-moved (TypeToBeMoved. :ok))
    (fn-to-be-moved (RecordToBeMoved. :ok))))